refactor code folder
