+++ lib/components/cardComponents/entryCard.dart
