+++ lib/core/alphabetState.dart
@@ -1,7 +1,7 @@
-import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
 
-///Stateful Aplphabet handler that is initialised for use by
+import 'package:flutter/material.dart';
+
+///Stateful Alphabet handler that is initialised for use by
 ///calling [Alphabet().createState()]
 ///The [createState()] adds an awareness of the widget state to
 ///the widget tree and enables overlaying the [AlphabetButton] widget
@@ -52,39 +52,5 @@ class _AlphabetState extends State<Alphabet> {
   ///reset the [newAlpha] to an empty [List<String>]
   void reset() {
     newAlpha = [];
-  }
-}
-
-///[AlphabetButton] creates a [Card] that is wrapped with
-///a [GestureDetector] it is usually used in conjunction with a [LetterMap]
-///and an [Alphabet] state handler
-class AlphabetButton extends StatelessWidget {
-  AlphabetButton({this.alphabet, this.active, this.onPressed});
-
-  /// [bool] active manages the state of the UI of the button
-  /// if the button has been triggered the button color is transformed
-  /// to a completely transparent [Card] with a minor elevation backdrop.
-  final bool active;
-
-  /// [String] alphabet: is the child of the [AlphabetButton]
-  /// the actual alphabet to be displayed
-  final String alphabet;
-
-  /// [Function] a function to be triggered on the button press detected.
-  final Function onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: onPressed,
-      child: Card(
-          child: Padding(
-            padding: const EdgeInsets.all(14.0),
-            child: Text(alphabet,
-                style: GoogleFonts.poppins(
-                    fontWeight: FontWeight.bold, fontSize: 17)),
-          ),
-          color: active ? Colors.white.withOpacity(.6) : Colors.transparent),
-    );
   }
 }
