+++ lib/core/utilities/dictionary.dart
