+++ lib/core/utilities/dictionaryActivity.dart
