+++ lib/core/utilities/entryHandler.dart
@@ -1,26 +1,32 @@
 import 'package:flutter/material.dart';
-import 'package:wordsmith/utilities/words.dart';
+import 'package:wordsmith/core/alphabetState.dart';
+import 'package:wordsmith/core/utilities/words.dart';
 import 'dictionaryActivity.dart';
 import 'scoreKeeper.dart';
-import 'package:wordsmith/components/displayComponents/buttons/alphabets.dart';
 
 class EntryHandler {
   ///The entryHandler acts as a binder for many gameplay relating functions
   ///The entryHandler handles thw word creation by calling the [WordGenerator]
   ///It handles the Alphabet Widgets across both single and multiplayer gameplays
   /// it also combines with an alphabetHandler [Alphabet]to control the state of alphabet
-  /// widgets 
+  /// widgets
   EntryHandler({this.wordGenerator});
+
   ///User Game Entries
   String entry;
+
   /// Game Parent Word from which the subWords are expected
   String gameWord;
+
   /// An EntryList formed from the entries converted to renderable widgets
   List<Widget> entryList = [];
+
   /// Alphabet Widgets
   final List<Widget> alphaWidgets = [];
+
   ///ScoreKeeper keeps track of the player validation count and attached score rendering
   ScoreKeeper scoreKeeper = ScoreKeeper();
+
   ///Alphabet state instance to handle alphabet widget states
   final alphabetHandler = Alphabet().createState();
   final Words wordGenerator;
@@ -29,10 +35,11 @@ class EntryHandler {
   List<String> getWord() {
     return wordGenerator.getRandom();
   }
+
   ///Validate the entered word,
   ///if the [returnScore] value is parsed then
   ///the [ScoreKeeper] saves the score of that entry
-  ///else it passes and just returns the [bool] outcome 
+  ///else it passes and just returns the [bool] outcome
   bool validate({String entry, @required bool returnScore}) {
     gameWord = wordGenerator.allAlphabets();
     bool validated = verifyWord(gameWord, entry);
@@ -43,10 +50,11 @@ class EntryHandler {
     }
     return validated;
   }
-  
+
   String getGameWord() {
     return wordGenerator.allAlphabets();
   }
+
   ///insert a value into the [entryList]
   insert(String entry) {
     bool correct = validate(entry: entry, returnScore: true);
