+++ lib/core/utilities/localData.dart
