+++ lib/core/utilities/words.dart
