+++ lib/main.dart
@@ -1,15 +1,16 @@
 import 'package:flutter/material.dart';
-import 'package:wordsmith/screens/friendsScreen.dart';
-import 'package:wordsmith/screens/levelSelectScreen.dart';
-import 'package:wordsmith/screens/playerScreen.dart';
-import 'package:wordsmith/screens/registerScreen.dart';
+import 'package:wordsmith/handlers/stateHandlers/providerHandlers/themeData.dart';
+import 'package:wordsmith/handlers/stateHandlers/providerHandlers/userData.dart';
+import 'package:wordsmith/screens/pages/friendsScreen.dart';
+import 'package:wordsmith/screens/pages/levelSelectScreen.dart';
+import 'package:wordsmith/screens/pages/loadingScreen.dart';
+import 'package:wordsmith/screens/pages/playerScreen.dart';
+import 'package:wordsmith/screens/pages/registerScreen.dart';
+import 'package:wordsmith/screens/pages/signInScreen.dart';
 import 'package:wordsmith/screens/singlePlayerLevels/singleLevelOne.dart';
 import 'package:wordsmith/screens/multiPlayerLevels/multiLevelOne.dart';
-import 'package:wordsmith/screens/loadingScreen.dart';
-import 'package:wordsmith/screens/signInScreen.dart';
 import 'package:provider/provider.dart';
-import 'package:wordsmith/userProvider/userData.dart';
-import 'package:wordsmith/userProvider/themeData.dart';
+
 
 void main() {
   runApp(MultiProvider(
@@ -25,10 +26,8 @@ void main() {
         'LoadingScreen': (context) => LoadingScreen(),
         'SingleLevelOne': (context) => SingleLevelOne(),
         'MultiLevelOne': (context) => MultiLevelOne(),
-
         'SignInPage': (context) => SignInPage(),
         'RegisterPage': (context) => RegisterScreen(),
-        // 'ChooseOpponent': (context) => ChooseOpponent(),
         'PlayerScreen': (context) => PlayerScreen(),
         'FriendScreen': (context) => FriendScreen()
       },
