+++ lib/screens/pages/friendsScreen.dart
@@ -2,10 +2,10 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import "package:flutter/material.dart";
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
-import 'package:wordsmith/components/displayComponents/card/userCard.dart';
-import 'package:wordsmith/userProvider/themeData.dart';
-import 'package:wordsmith/userProvider/userData.dart';
-import 'package:wordsmith/utilities/constants.dart';
+import 'package:wordsmith/components/cardComponents/userCard.dart';
+import 'package:wordsmith/core/utilities/constants.dart';
+import 'package:wordsmith/handlers/stateHandlers/providerHandlers/themeData.dart';
+import 'package:wordsmith/handlers/stateHandlers/providerHandlers/userData.dart';
 
 class FriendScreen extends StatefulWidget {
   @override
@@ -28,7 +28,7 @@ class _FriendScreenState extends State<FriendScreen> {
   @override
   Widget build(BuildContext context) {
     final Data userData = Provider.of<Data>(context);
-    final theme = Provider.of<AppThemeData>(context);
+    final AppThemeData theme = Provider.of<AppThemeData>(context);
     double height = MediaQuery.of(context).size.height;
     double width = MediaQuery.of(context).size.width;
