+++ lib/screens/pages/signInScreen.dart
@@ -1,15 +1,19 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'dart:ui' as ui;
-import 'package:wordsmith/components/displayComponents/inputFields/inputField.dart';
-import 'package:wordsmith/components/displayComponents/logo.dart';
-import 'package:wordsmith/userProvider/themeData.dart';
-import 'package:wordsmith/utilities/constants.dart';
+
+import 'package:wordsmith/components/inputComponents/buttons/slimButtons.dart';
+import 'package:wordsmith/components/inputComponents/textFields/inputField.dart';
+import 'package:wordsmith/core/logo.dart';
+import 'package:wordsmith/core/utilities/constants.dart';
+
+
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:provider/provider.dart';
-import 'package:wordsmith/userProvider/userData.dart';
-import 'package:wordsmith/components/displayComponents/buttons/slimButtons.dart';
+
 import 'package:modal_progress_hud/modal_progress_hud.dart';
+import 'package:wordsmith/handlers/stateHandlers/providerHandlers/themeData.dart';
+import 'package:wordsmith/handlers/stateHandlers/providerHandlers/userData.dart';
 
 class SignInPage extends StatefulWidget {
   @override
@@ -139,8 +143,8 @@ class _SignInPageState extends State<SignInPage> with TickerProviderStateMixin {
                                   .collection('users')
                                   .document(loggedInUserId)
                                   .get()
-                                  .then((value) =>
-                                      userData.updateUserName(value['username']));
+                                  .then((value) => userData
+                                      .updateUserName(value['username']));
                               Navigator.pushReplacementNamed(
                                   context, 'PlayerScreen');
                             }
