+++ lib/screens/popUps/dialogs/waitingForOpponent.dart
@@ -2,9 +2,10 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
+import 'package:wordsmith/core/utilities/constants.dart';
+import 'package:wordsmith/handlers/stateHandlers/providerHandlers/userData.dart';
+import 'package:wordsmith/handlers/stateHandlers/streamLogic/waitingGameStream.dart';
 import 'package:wordsmith/screens/multiPlayerLevels/multiLevelOne.dart';
-import 'package:wordsmith/userProvider/userData.dart';
-import 'package:wordsmith/utilities/constants.dart';
 
 class WaitingForOpponent extends StatelessWidget {
   ///A popup streambuilder widget that listens for changes
@@ -91,47 +92,3 @@ class WaitingForOpponent extends StatelessWidget {
   }
 }
 
-class GameStream extends StatelessWidget {
-  ///Stream Listener that carries out the [WaitingForOpponent] functionality
-  const GameStream({
-    Key key,
-    @required this.firestore,
-    @required this.userID,
-    @required this.userData,
-    this.opponentName,
-  }) : super(key: key);
-
-  final Firestore firestore;
-  final String userID;
-  final Data userData;
-  final String opponentName;
-
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder<DocumentSnapshot>(
-        stream: firestore.collection('users').document(userID).snapshots(),
-        builder: (context, snapshot) {
-          if (snapshot.hasData) {
-            List activeGames = snapshot.data['acceptedChallenges'];
-            print(activeGames);
-            if (activeGames.contains(userData.currentUserID)) {
-              WidgetsBinding.instance.addPostFrameCallback(
-                (_) {
-                  Navigator.pushReplacement(
-                    context,
-                    MaterialPageRoute(builder: (context) {
-                      return MultiLevelOne(
-                          opponentName: opponentName,
-                          opponentID: userID,
-                          randomIndex: 1);
-                    }),
-                  );
-                },
-              );
-            }
-            return Container();
-          }
-          return Container();
-        });
-  }
-}
